+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v1.1.0
+- Added a title screen with a demo mode.
 - Added transparency to the logo image.
 
 v1.0.0